CS Gas possible fix
The duration effect code is passing negative (or zero) values to the timer , creating crash on the effect. 

https://github.com/eezstreet/SWATEliteForce/blob/52ff0e3158ff1842a5ea36e65e73ecfe79bd9aba/Source/Game/SwatGame/Classes/SwatPlayer.uc#L2535

DistanceEffect = (600 - Distance)/(600); 

The radius must be less than 600 to avoid crash.
So far it worked in my tests.

On the contrary , making radius over 600 makes constant crashes while player is outside the 600 distance.